novas atualizações

+++ CadastroServer/src/cadastroserver/controller/ProdutoJpaController.java
@@ -1,88 +1,42 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package cadastroserver.controller;
 
 import cadastroserver.controller.exceptions.NonexistentEntityException;
 import cadastroserver.controller.exceptions.PreexistingEntityException;
-import java.io.Serializable;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
+import cadastroserver.model.Produto;
+import cadastroserver.model.MovimentoCompra;
+import cadastroserver.model.MovimentoVenda;
+
+import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
-import cadastroserver.model.MovimentoVenda;
-import java.util.ArrayList;
-import java.util.Collection;
-import cadastroserver.model.MovimentoCompra;
-import cadastroserver.model.Produto;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 
-/**
- *
- * @author Madu
- */
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.util.List;
+import java.util.ArrayList;
+
 public class ProdutoJpaController implements Serializable {
+
+    private final EntityManagerFactory emf;
 
     public ProdutoJpaController(EntityManagerFactory emf) {
         this.emf = emf;
     }
-    private EntityManagerFactory emf = null;
 
-    public EntityManager getEntityManager() {
+    private EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
     public void create(Produto produto) throws PreexistingEntityException, Exception {
-        if (produto.getMovimentoVendaCollection() == null) {
-            produto.setMovimentoVendaCollection(new ArrayList<MovimentoVenda>());
-        }
-        if (produto.getMovimentoCompraCollection() == null) {
-            produto.setMovimentoCompraCollection(new ArrayList<MovimentoCompra>());
-        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            Collection<MovimentoVenda> attachedMovimentoVendaCollection = new ArrayList<MovimentoVenda>();
-            for (MovimentoVenda movimentoVendaCollectionMovimentoVendaToAttach : produto.getMovimentoVendaCollection()) {
-                movimentoVendaCollectionMovimentoVendaToAttach = em.getReference(movimentoVendaCollectionMovimentoVendaToAttach.getClass(), movimentoVendaCollectionMovimentoVendaToAttach.getIdMovimentoVenda());
-                attachedMovimentoVendaCollection.add(movimentoVendaCollectionMovimentoVendaToAttach);
-            }
-            produto.setMovimentoVendaCollection(attachedMovimentoVendaCollection);
-            Collection<MovimentoCompra> attachedMovimentoCompraCollection = new ArrayList<MovimentoCompra>();
-            for (MovimentoCompra movimentoCompraCollectionMovimentoCompraToAttach : produto.getMovimentoCompraCollection()) {
-                movimentoCompraCollectionMovimentoCompraToAttach = em.getReference(movimentoCompraCollectionMovimentoCompraToAttach.getClass(), movimentoCompraCollectionMovimentoCompraToAttach.getIdMovimentoCompra());
-                attachedMovimentoCompraCollection.add(movimentoCompraCollectionMovimentoCompraToAttach);
-            }
-            produto.setMovimentoCompraCollection(attachedMovimentoCompraCollection);
             em.persist(produto);
-            for (MovimentoVenda movimentoVendaCollectionMovimentoVenda : produto.getMovimentoVendaCollection()) {
-                Produto oldIdProdutoOfMovimentoVendaCollectionMovimentoVenda = movimentoVendaCollectionMovimentoVenda.getIdProduto();
-                movimentoVendaCollectionMovimentoVenda.setIdProduto(produto);
-                movimentoVendaCollectionMovimentoVenda = em.merge(movimentoVendaCollectionMovimentoVenda);
-                if (oldIdProdutoOfMovimentoVendaCollectionMovimentoVenda != null) {
-                    oldIdProdutoOfMovimentoVendaCollectionMovimentoVenda.getMovimentoVendaCollection().remove(movimentoVendaCollectionMovimentoVenda);
-                    oldIdProdutoOfMovimentoVendaCollectionMovimentoVenda = em.merge(oldIdProdutoOfMovimentoVendaCollectionMovimentoVenda);
-                }
-            }
-            for (MovimentoCompra movimentoCompraCollectionMovimentoCompra : produto.getMovimentoCompraCollection()) {
-                Produto oldIdProdutoOfMovimentoCompraCollectionMovimentoCompra = movimentoCompraCollectionMovimentoCompra.getIdProduto();
-                movimentoCompraCollectionMovimentoCompra.setIdProduto(produto);
-                movimentoCompraCollectionMovimentoCompra = em.merge(movimentoCompraCollectionMovimentoCompra);
-                if (oldIdProdutoOfMovimentoCompraCollectionMovimentoCompra != null) {
-                    oldIdProdutoOfMovimentoCompraCollectionMovimentoCompra.getMovimentoCompraCollection().remove(movimentoCompraCollectionMovimentoCompra);
-                    oldIdProdutoOfMovimentoCompraCollectionMovimentoCompra = em.merge(oldIdProdutoOfMovimentoCompraCollectionMovimentoCompra);
-                }
-            }
             em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findProduto(produto.getIdProduto()) != null) {
-                throw new PreexistingEntityException("Produto " + produto + " already exists.", ex);
-            }
-            throw ex;
+        } catch (EntityExistsException e) {
+            throw new PreexistingEntityException("Produto " + produto + " already exists.", e);
         } finally {
             if (em != null) {
                 em.close();
@@ -95,70 +49,10 @@ public class ProdutoJpaController implements Serializable {
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            Produto persistentProduto = em.find(Produto.class, produto.getIdProduto());
-            Collection<MovimentoVenda> movimentoVendaCollectionOld = persistentProduto.getMovimentoVendaCollection();
-            Collection<MovimentoVenda> movimentoVendaCollectionNew = produto.getMovimentoVendaCollection();
-            Collection<MovimentoCompra> movimentoCompraCollectionOld = persistentProduto.getMovimentoCompraCollection();
-            Collection<MovimentoCompra> movimentoCompraCollectionNew = produto.getMovimentoCompraCollection();
-            Collection<MovimentoVenda> attachedMovimentoVendaCollectionNew = new ArrayList<MovimentoVenda>();
-            for (MovimentoVenda movimentoVendaCollectionNewMovimentoVendaToAttach : movimentoVendaCollectionNew) {
-                movimentoVendaCollectionNewMovimentoVendaToAttach = em.getReference(movimentoVendaCollectionNewMovimentoVendaToAttach.getClass(), movimentoVendaCollectionNewMovimentoVendaToAttach.getIdMovimentoVenda());
-                attachedMovimentoVendaCollectionNew.add(movimentoVendaCollectionNewMovimentoVendaToAttach);
-            }
-            movimentoVendaCollectionNew = attachedMovimentoVendaCollectionNew;
-            produto.setMovimentoVendaCollection(movimentoVendaCollectionNew);
-            Collection<MovimentoCompra> attachedMovimentoCompraCollectionNew = new ArrayList<MovimentoCompra>();
-            for (MovimentoCompra movimentoCompraCollectionNewMovimentoCompraToAttach : movimentoCompraCollectionNew) {
-                movimentoCompraCollectionNewMovimentoCompraToAttach = em.getReference(movimentoCompraCollectionNewMovimentoCompraToAttach.getClass(), movimentoCompraCollectionNewMovimentoCompraToAttach.getIdMovimentoCompra());
-                attachedMovimentoCompraCollectionNew.add(movimentoCompraCollectionNewMovimentoCompraToAttach);
-            }
-            movimentoCompraCollectionNew = attachedMovimentoCompraCollectionNew;
-            produto.setMovimentoCompraCollection(movimentoCompraCollectionNew);
-            produto = em.merge(produto);
-            for (MovimentoVenda movimentoVendaCollectionOldMovimentoVenda : movimentoVendaCollectionOld) {
-                if (!movimentoVendaCollectionNew.contains(movimentoVendaCollectionOldMovimentoVenda)) {
-                    movimentoVendaCollectionOldMovimentoVenda.setIdProduto(null);
-                    movimentoVendaCollectionOldMovimentoVenda = em.merge(movimentoVendaCollectionOldMovimentoVenda);
-                }
-            }
-            for (MovimentoVenda movimentoVendaCollectionNewMovimentoVenda : movimentoVendaCollectionNew) {
-                if (!movimentoVendaCollectionOld.contains(movimentoVendaCollectionNewMovimentoVenda)) {
-                    Produto oldIdProdutoOfMovimentoVendaCollectionNewMovimentoVenda = movimentoVendaCollectionNewMovimentoVenda.getIdProduto();
-                    movimentoVendaCollectionNewMovimentoVenda.setIdProduto(produto);
-                    movimentoVendaCollectionNewMovimentoVenda = em.merge(movimentoVendaCollectionNewMovimentoVenda);
-                    if (oldIdProdutoOfMovimentoVendaCollectionNewMovimentoVenda != null && !oldIdProdutoOfMovimentoVendaCollectionNewMovimentoVenda.equals(produto)) {
-                        oldIdProdutoOfMovimentoVendaCollectionNewMovimentoVenda.getMovimentoVendaCollection().remove(movimentoVendaCollectionNewMovimentoVenda);
-                        oldIdProdutoOfMovimentoVendaCollectionNewMovimentoVenda = em.merge(oldIdProdutoOfMovimentoVendaCollectionNewMovimentoVenda);
-                    }
-                }
-            }
-            for (MovimentoCompra movimentoCompraCollectionOldMovimentoCompra : movimentoCompraCollectionOld) {
-                if (!movimentoCompraCollectionNew.contains(movimentoCompraCollectionOldMovimentoCompra)) {
-                    movimentoCompraCollectionOldMovimentoCompra.setIdProduto(null);
-                    movimentoCompraCollectionOldMovimentoCompra = em.merge(movimentoCompraCollectionOldMovimentoCompra);
-                }
-            }
-            for (MovimentoCompra movimentoCompraCollectionNewMovimentoCompra : movimentoCompraCollectionNew) {
-                if (!movimentoCompraCollectionOld.contains(movimentoCompraCollectionNewMovimentoCompra)) {
-                    Produto oldIdProdutoOfMovimentoCompraCollectionNewMovimentoCompra = movimentoCompraCollectionNewMovimentoCompra.getIdProduto();
-                    movimentoCompraCollectionNewMovimentoCompra.setIdProduto(produto);
-                    movimentoCompraCollectionNewMovimentoCompra = em.merge(movimentoCompraCollectionNewMovimentoCompra);
-                    if (oldIdProdutoOfMovimentoCompraCollectionNewMovimentoCompra != null && !oldIdProdutoOfMovimentoCompraCollectionNewMovimentoCompra.equals(produto)) {
-                        oldIdProdutoOfMovimentoCompraCollectionNewMovimentoCompra.getMovimentoCompraCollection().remove(movimentoCompraCollectionNewMovimentoCompra);
-                        oldIdProdutoOfMovimentoCompraCollectionNewMovimentoCompra = em.merge(oldIdProdutoOfMovimentoCompraCollectionNewMovimentoCompra);
-                    }
-                }
-            }
+            em.merge(produto);
             em.getTransaction().commit();
-        } catch (Exception ex) {
-            String msg = ex.getLocalizedMessage();
-            if (msg == null || msg.length() == 0) {
-                Integer id = produto.getIdProduto();
-                if (findProduto(id) == null) {
-                    throw new NonexistentEntityException("The produto with id " + id + " no longer exists.");
-                }
-            }
-            throw ex;
+        } catch (EntityNotFoundException e) {
+            throw new NonexistentEntityException("The produto with id " + produto.getIdProduto() + " no longer exists.", e);
         } finally {
             if (em != null) {
                 em.close();
@@ -171,22 +65,9 @@ public class ProdutoJpaController implements Serializable {
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            Produto produto;
-            try {
-                produto = em.getReference(Produto.class, id);
-                produto.getIdProduto();
-            } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The produto with id " + id + " no longer exists.", enfe);
-            }
-            Collection<MovimentoVenda> movimentoVendaCollection = produto.getMovimentoVendaCollection();
-            for (MovimentoVenda movimentoVendaCollectionMovimentoVenda : movimentoVendaCollection) {
-                movimentoVendaCollectionMovimentoVenda.setIdProduto(null);
-                movimentoVendaCollectionMovimentoVenda = em.merge(movimentoVendaCollectionMovimentoVenda);
-            }
-            Collection<MovimentoCompra> movimentoCompraCollection = produto.getMovimentoCompraCollection();
-            for (MovimentoCompra movimentoCompraCollectionMovimentoCompra : movimentoCompraCollection) {
-                movimentoCompraCollectionMovimentoCompra.setIdProduto(null);
-                movimentoCompraCollectionMovimentoCompra = em.merge(movimentoCompraCollectionMovimentoCompra);
+            Produto produto = em.find(Produto.class, id);
+            if (produto == null) {
+                throw new NonexistentEntityException("The produto with id " + id + " no longer exists.");
             }
             em.remove(produto);
             em.getTransaction().commit();
@@ -208,9 +89,10 @@ public class ProdutoJpaController implements Serializable {
     private List<Produto> findProdutoEntities(boolean all, int maxResults, int firstResult) {
         EntityManager em = getEntityManager();
         try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            CriteriaBuilder cb = em.getCriteriaBuilder();
+            CriteriaQuery<Produto> cq = cb.createQuery(Produto.class);
             cq.select(cq.from(Produto.class));
-            Query q = em.createQuery(cq);
+            TypedQuery<Produto> q = em.createQuery(cq);
             if (!all) {
                 q.setMaxResults(maxResults);
                 q.setFirstResult(firstResult);
@@ -233,14 +115,15 @@ public class ProdutoJpaController implements Serializable {
     public int getProdutoCount() {
         EntityManager em = getEntityManager();
         try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            CriteriaBuilder cb = em.getCriteriaBuilder();
+            CriteriaQuery<Long> cq = cb.createQuery(Long.class);
             Root<Produto> rt = cq.from(Produto.class);
-            cq.select(em.getCriteriaBuilder().count(rt));
-            Query q = em.createQuery(cq);
-            return ((Long) q.getSingleResult()).intValue();
+            cq.select(cb.count(rt));
+            TypedQuery<Long> q = em.createQuery(cq);
+            return q.getSingleResult().intValue();
         } finally {
             em.close();
         }
     }
-    
+
 }
